namespace w3c proof
